pyproject.toml
[project]
name = "tabbycat-manager2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coloredlogs>=15.0.1",
    "cryptography>=44.0.2",
    "flet[all]==0.27.5",
    "google-api-python-client>=2.166.0",
    "pandas>=2.2.3",
    "tabbycat-api-2",
]

[tool.flet]
app.module = "main"

[tool.uv.sources]
tabbycat-api-2 = { git = "https://github.com/inohan/tabbycat-api-2.git" }
